+++ .assets/provision/install_pwsh.sh
@@ -3,23 +3,19 @@
 sudo .assets/provision/install_pwsh.sh >/dev/null
 '
 if [ $EUID -ne 0 ]; then
-  printf '\e[31;1mRun the script as root.\e[0m\n'
+  printf '\e[31;1mRun the script as root.\e[0m\n' >&2
   exit 1
 fi
 
+# dotsource file with common functions
+. .assets/provision/source.sh
+
+# define variables
 APP='pwsh'
 REL=$1
 retry_count=0
-# try 10 times to get latest release if not provided as a parameter
-while [ -z "$REL" ]; do
-  REL=$(curl -sk https://api.github.com/repos/PowerShell/PowerShell/releases/latest | sed -En 's/.*"tag_name": "v?([^"]*)".*/\1/p')
-  ((retry_count++))
-  if [ $retry_count -eq 10 ]; then
-    printf "\e[33m$APP version couldn't be retrieved\e[0m\n" >&2
-    exit 0
-  fi
-  [[ "$REL" =~ ^v?[0-9]+\.[0-9]+\.[0-9]+$ ]] || echo 'retrying...' >&2
-done
+# get latest release if not provided as a parameter
+[ -z "$REL" ] && REL="$(get_gh_release_latest --owner 'PowerShell' --repo 'PowerShell')"
 # return latest release
 echo $REL
 
@@ -39,15 +35,19 @@ case $SYS_ID in
 alpine)
   apk add --no-cache ncurses-terminfo-base krb5-libs libgcc libintl libssl1.1 libstdc++ tzdata userspace-rcu zlib icu-libs >&2 2>/dev/null
   apk -X https://dl-cdn.alpinelinux.org/alpine/edge/main add --no-cache lttng-ust >&2 2>/dev/null
+  # create temporary dir for the downloaded binary
   TMP_DIR=$(mktemp -dp "$PWD")
-  retry_count=0
-  while [[ ! -f "$TMP_DIR/$APP.tar.gz" && $retry_count -lt 10 ]]; do
-    curl -#Lko "$TMP_DIR/$APP.tar.gz" "https://github.com/PowerShell/PowerShell/releases/download/v${REL}/powershell-${REL}-linux-alpine-x64.tar.gz"
-    ((retry_count++))
-  done
-  mkdir -p /opt/microsoft/powershell/7 && tar -zxf "$TMP_DIR/$APP.tar.gz" -C /opt/microsoft/powershell/7
+  # calculate download uri
+  URL="https://github.com/PowerShell/PowerShell/releases/download/v${REL}/powershell-${REL}-linux-alpine-x64.tar.gz"
+  # download and install file
+  if download_file --uri $URL --target_dir $TMP_DIR; then
+    mkdir -p /opt/microsoft/powershell/7
+    tar -zxf "$TMP_DIR/$(basename $URL)" -C /opt/microsoft/powershell/7
+    chmod +x /opt/microsoft/powershell/7/pwsh
+    ln -s /opt/microsoft/powershell/7/pwsh /usr/bin/pwsh
+  fi
+  # remove temporary dir
   rm -fr "$TMP_DIR"
-  chmod +x /opt/microsoft/powershell/7/pwsh && ln -s /opt/microsoft/powershell/7/pwsh /usr/bin/pwsh
   ;;
 arch)
   if pacman -Qqe paru &>/dev/null; then
@@ -76,13 +76,15 @@ debian | ubuntu)
   else
     export DEBIAN_FRONTEND=noninteractive
     [ "$SYS_ID" = 'debian' ] && apt-get update >&2 && apt-get install -y libicu67 >&2 2>/dev/null || true
+    # create temporary dir for the downloaded binary
     TMP_DIR=$(mktemp -dp "$PWD")
-    retry_count=0
-    while [[ ! -f "$TMP_DIR/$APP.deb" && $retry_count -lt 10 ]]; do
-      curl -#Lko "$TMP_DIR/$APP.deb" "https://github.com/PowerShell/PowerShell/releases/download/v${REL}/powershell_${REL}-1.deb_amd64.deb"
-      ((retry_count++))
-    done
-    dpkg -i "$TMP_DIR/$APP.deb" >&2 2>/dev/null || binary=true
+    # calculate download uri
+    URL="https://github.com/PowerShell/PowerShell/releases/download/v${REL}/powershell_${REL}-1.deb_amd64.deb"
+    # download and install file
+    if download_file --uri $URL --target_dir $TMP_DIR; then
+      dpkg -i "$TMP_DIR/$(basename $URL)" >&2 2>/dev/null || binary=true
+    fi
+    # remove temporary dir
     rm -fr "$TMP_DIR"
   fi
   ;;
@@ -91,18 +93,20 @@ debian | ubuntu)
   ;;
 esac
 
-if [ "$binary" = true ]; then
+if [ "$binary" = true ] && [ -n "$REL" ]; then
   echo 'Installing from binary.' >&2
   [ "$SYS_ID" = 'opensuse' ] && zypper in -y libicu >&2 2>/dev/null || true
+  # create temporary dir for the downloaded binary
   TMP_DIR=$(mktemp -dp "$PWD")
-  retry_count=0
-  while [[ ! -f "$TMP_DIR/$APP.tar.gz" && $retry_count -lt 10 ]]; do
-    curl -#Lko "$TMP_DIR/$APP.tar.gz" "https://github.com/PowerShell/PowerShell/releases/download/v${REL}/powershell-${REL}-linux-x64.tar.gz"
-    ((retry_count++))
-  done
-  mkdir -p /opt/microsoft/powershell/7
-  tar -zxf "$TMP_DIR/$APP.tar.gz" -C /opt/microsoft/powershell/7
+  # calculate download uri
+  URL="https://github.com/PowerShell/PowerShell/releases/download/v${REL}/powershell-${REL}-linux-x64.tar.gz"
+  # download and install file
+  if download_file --uri $URL --target_dir $TMP_DIR; then
+    mkdir -p /opt/microsoft/powershell/7
+    tar -zxf "$TMP_DIR/$(basename $URL)" -C /opt/microsoft/powershell/7
+    chmod +x /opt/microsoft/powershell/7/pwsh
+    [ -f /usr/bin/pwsh ] || ln -s /opt/microsoft/powershell/7/pwsh /usr/bin/pwsh
+  fi
+  # remove temporary dir
   rm -fr "$TMP_DIR"
-  chmod +x /opt/microsoft/powershell/7/pwsh
-  [ -f /usr/bin/pwsh ] || ln -s /opt/microsoft/powershell/7/pwsh /usr/bin/pwsh
 fi
