+++ .assets/provision/install_ripgrep.sh
@@ -3,7 +3,7 @@
 sudo .assets/provision/install_ripgrep.sh >/dev/null
 '
 if [ $EUID -ne 0 ]; then
-  printf '\e[31;1mRun the script as root.\e[0m\n'
+  printf '\e[31;1mRun the script as root.\e[0m\n' >&2
   exit 1
 fi
 
@@ -26,18 +26,14 @@ debian | ubuntu)
   ;;
 esac
 
+# dotsource file with common functions
+. .assets/provision/source.sh
+
+# define variables
 REL=$1
 retry_count=0
-# try 10 times to get latest release if not provided as a parameter
-while [ -z "$REL" ]; do
-  REL=$(curl -sk https://api.github.com/repos/BurntSushi/ripgrep/releases/latest | sed -En 's/.*"tag_name": "v?([^"]*)".*/\1/p')
-  ((retry_count++))
-  if [ $retry_count -eq 10 ]; then
-    printf "\e[33m$APP version couldn't be retrieved\e[0m\n" >&2
-    exit 0
-  fi
-  [[ "$REL" =~ ^v?[0-9]+\.[0-9]+\.[0-9]+$ ]] || echo 'retrying...' >&2
-done
+# get latest release if not provided as a parameter
+[ -z "$REL" ] && REL="$(get_gh_release_latest --owner 'BurntSushi' --repo 'ripgrep')"
 # return latest release
 echo $REL
 
@@ -72,16 +68,19 @@ opensuse)
   ;;
 esac
 
-if [ "$binary" = true ]; then
+if [ "$binary" = true ] && [ -n "$REL" ]; then
   echo 'Installing from binary.' >&2
+  # create temporary dir for the downloaded binary
   TMP_DIR=$(mktemp -dp "$PWD")
-  retry_count=0
-  while [[ ! -f "$TMP_DIR/rg" && $retry_count -lt 10 ]]; do
-    curl -#Lk "https://github.com/BurntSushi/ripgrep/releases/download/${REL}/ripgrep-${REL}-x86_64-unknown-linux-musl.tar.gz" | tar -zx --strip-components=1 -C "$TMP_DIR"
-    ((retry_count++))
-  done
-  install -m 0755 "$TMP_DIR/rg" /usr/bin/
-  install -m 0644 "$TMP_DIR/doc/rg.1" "$(manpath | cut -d : -f 1)/man1/"
-  install -m 0644 "$TMP_DIR/complete/rg.bash" /etc/bash_completion.d/
+  # calculate download uri
+  URL="https://github.com/BurntSushi/ripgrep/releases/download/${REL}/ripgrep-${REL}-aarch64-unknown-linux-gnu.tar.gz"
+  # download and install file
+  if download_file --uri $URL --target_dir $TMP_DIR; then
+    tar -zxf "$TMP_DIR/$(basename $URL)" --strip-components=1 -C "$TMP_DIR"
+    install -m 0755 "$TMP_DIR/rg" /usr/bin/
+    install -m 0644 "$TMP_DIR/doc/rg.1" "$(manpath | cut -d : -f 1)/man1/"
+    install -m 0644 "$TMP_DIR/complete/rg.bash" /etc/bash_completion.d/
+  fi
+  # remove temporary dir
   rm -fr "$TMP_DIR"
 fi
